+++ CHANGELOG.md
@@ -59,6 +59,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
     annotations.
 -   Fix compliance with ISO 14496-1:2010 when writing AAC Codec Config OBUs
     (AOMediaCodec/libiamf#119).
+-   Permit one fully trimmed audio frame at the end of a substream.
 
 ## [1.0.0] - 2024-01-26